Tried fixing incoherent behavior.
The absence of off for pitch 40 was "fixed" by pushing the off event on the score before the note event at the same dt. This seems strange, as events within the same dt should not obfuscate each other. This probably has to do with how the "nextEvents" vector is constructed in Renderer.h (simply by pulling from the model's events).

The non-association of any end-of-note with the release of key 62 (that should trigger a !40 but does not at the end of the test) has not been fixed.

Exceptions have been added in Renderer.h :

- One checks that the events registered in the map are not start events
- The other replaces the simple if/else checking if map.find(key) does not return map.end()

Planting these exceptions has revealed that the two subsequent uses of key 62 are indeed erroneous. They have been replaced with 62 and 61. 

Trying to restrict construction of nextEvents by using a while(nextEvents.empty() || Events::hasStart<Model>(nextEvents)) leads to even more incorrect results in the coherence test (and an infinite loop in the robustness test). Therefore, the cause of the other incoherent behavior cannot be found.