Fixed tests to account for proper behavior
dt=0 denotes SYNCHRONOUS events with the previous non-zero dt. THIS IS WHAT NEEDS TO BE DONE to synchronize events. This was the major change needed for the test to work.

However, there was still incoherent behavior to fix : because renderer.nextEvents relies on a pull, it emptied the modelEvents chronology when the last relevant keypress (the last keypress with an ON event associated) occurred. This is now fixed with a boolean (lastEventPulled) ; the renderer indicates that it still has events available if lastEventPulled is true, so that other commands can still go through until the relevant release occurs. 

This correction allows us to see that the paradigm is such that synchronous events CANNOT be artificially desynchronized by the player. Only endings synchronized with beginnings can be desynchronized. 

To evidence this, test prints were emphasized and improved.

However, one problem remains. With the "coherent" partition found to be incomplete, I intended to test the system with a partition that takes advantage of all commands (since the last inputs in the incomplete test are useless). 

What this reveals is an inconspicuous non-ending to a note of pitch 70, while a note of pitch 75 is correctly ended. So there are still issues to solve.