Corrected non-ending to note 70 in full partition test
In more detail, this bug occurred because the !70 was pushed, isolated, and not immediately following the 70. It was thus never registered as the corresponding ending, and never triggered. 

To fix this, it is necessary to keep track of all "incomplete" events. This was done by creating a private list (it is a list so as to allow removal without shifting) of such events, containing both a copy of the events in question and a reference to the empty set that follows so it can be filled with the endings once they appear.

The check for a corresponding event, the construction of an insert set (which was already described in pushEvent), and the process of checking if an empty set can be made into an insert set when there are incomplete events pending have all been factorized into private methods.

As a bonus, all relevant methods in the Event header have now been marked as const where appropriate. 

NOTE : this correction does not guarantee that such an occurrence is entirely prevented ; further testing needed.