Test against chord release desynchronization
This new test reveals an incoherence in how the model chronology operates : it should consist of a strictly start/end/start/end-shaped sequence. This works in the case of the displaced ending partition, but for some reason, not with this new test.

The merging function is aimed at avoiding this two-ending-sets situation, but it seems to fail.

So in the case of this happening, the code was updated in such a way that : 

- The renderer keeps a list of note endings that have been orphaned (met with a key press and left aside)
- It pulls in a fifo manner from that list when a release is associated with an empty set.