After consultation, created option for event completion
Event completion apparently is NOT SUPPOSED to happen in the model as it is so far defined. Ending events should be merged into clusters and not displaced forward EVEN IF releases happen.

Why ? I don't understand, but as this is apparently the specification, a new boolean field has been added to the Chronology class that toggles event completion.

For now, the Renderer constructs its Model chronology with that boolean set to true. Further discussion will be needed on the matter.