Chronology parameter struct, temporal resolution, detachEndings and more
This is a large commit that should have been split into many.

MAJOR :

1. detachEndings method

Chronology objects now have a detachEndings method that, when a start
event and its corresponding end are in the same set, separates them
as the chronology is finalized. This is used to prevent "shadow notes"
which are a common phenomenon in MIDI files for an unknown reason.

2. Temporal resolution

A Chronology can now be set to merge all events which dt is less than a
set threshold, rather than only merge events with a dt of 0.
More experiments need to be done on this matter ; the "correct" value
has not yet been found.

3. ChronologyParams::parameters and its presets

A namespace called ChronologyParams now hosts a type named "parameters",
which is a struct used to pass all parameters used in the Chronology
constructor (a la JavaScript). It must be initialized LIKE a struct,
as it was purposely left without constructors ; this is so the values
are NAMED for better clarity.

Presets have been made available : one for a general-purpose,
unmeet-enabled Chronology, and another one with unmeet turned off.
Both offer a temporal resolution of 0. This means they go unused
for the moment.

Renderers and MFPRenderers can use this same exact struct as an argument
to their constructors to set up their inner chronology.

------------------------------------------------------------------------

MINOR :

1. Moved mergeSets() and isMatchingEnd() to the Events.h file.
It was more semantically appropriate.

2. Made the Chronology type "iterable" ; one can run an enhanced 
for-loop
over it. Its begin() and end() iterators are simply those of its fifo.

3. Added the Renderer.getPartition() method.

4. Improved the display (via << ) of MFPEvents, so that they display
not only their MIDI note ID (e.g. 76) but the corresponding note
(in this case, E5).