Changed lib folder structure, modified ending transformation
To facilitate compilation, separate "cpp" and "include" folders have 
been created. In the process, erroneous specialization was fixed in the 
begin() and end() methods of the Chronology class that used noteData 
instead of the template T type, which was only possible because this 
separation wasn't properly done !

The "detachEndings" algorithm from the Chronology class was replaced 
with "shiftSameEventEndings", which shifts these endings to the head of 
the set. This is because they may not actually be the endings of the 
starting events they are synchronized with, but of identical events 
earlier in the partition. This means that these starting events actually 
have their own endings later on, and fixes the hasty endings of some 
notes.

To accomplish this, Events::mergeSets now has an option to merge at the 
beginning of a set and not the end. Similar options should be provided 
soon (like a way to "correspond" with a note only by pitch and not 
channel).