Added an overload for correspond and isMatchingEnd and changed shift logic
correspond and isMatchingEnd can now take a third argument of type 
"correspondOption", which is an enum class. The enumerators are only 
accessible through the specialization file. 

There are three enumerators : PITCH_AND_CHANNEL (the default, since it 
is 0 ; note that the default params for a Chronology now DO NOT 
INITIALIZE its shiftMode, using this default value instead), PITCH_ONLY 
and NONE. 

shiftSameEventEndings is now systematically called, but uses 
isMatchingEnd with the provided shiftMode.

The method's logic was also rewritten because its previous version 
erroneously caused events to be discarded from sets, resulting in 
missing notes - which is a rather critical bug.