Removed IME exception and fixed note stealing issue
Because the IME case was found to be a false positive in any case it was
fired (since it was always a single-controller case where no event has
been found, and these could be caused by trivial situations like
starting the program with keys already held down), the exception was
removed entirely.

However, to compensate, if a key already has a set of associated release
events, these are triggered with the key press. It is debatable whether
this is the most intuitive choice (rather than triggering them together
with the key release).

As well, voice stealing (the phenomenon in which two keys share one or
more pitches on the same channel and one will prematurely end the
other's events) has been addressed with a separate map transferring any
ending events to the most recent key associated with their beginnings.

These solutions have been tested in the score binding and seem to work.
Their effect on the MFP's complexity and expressiveness as a TIMS should
be evaluated formally to ensure they do not cause a loss similar to the
deactivation of the unmeet transformation.