Improved remove algorithms, renaming and commenting
"performVoiceStealing" was renamed to "preventVoiceStealing" for rather 
obvious reasons.

checkForEventCompletion now makes use of the std::erase_if C++20 method. 
Unfortunately, we cannot do the same for constructInsert set or the 
LastNoteOffWins's class preventVoiceStealing method, because they do 
more than just erase : they also require other operations on entries 
that don't satisfy the removal predicate. According to J.M. Celerier, 
there is no streamlined solution for this case, so we have to resort to 
old-fashioned iterator while loops.

However, we still gain something by replacing inserts of note-offs with 
the concatenation of a big note-off vector at the beginning of the note 
vector : less shifting overhead for one, and more readable code for two 
(getting read of the begin() + i syntax).

With this, the branch should be ready to merge.