First steps in the actual DLL API
The compilation chain under Linux with the mingw toolchain simply does not work. Environment switched to Windows (more precisely, Visual Studio) : the compiler specifications in the

With this, and the dummy function confirmed to work in Unity, the true prototype and code of pushMPTKEvent have been added. The renderer is kept solely on the C++ side, constructed from there and hidden from the C# code.

For the same reason, the call to performer.finalize() is hidden inside a top-level API function.

The parameters are inherited from the good old score implementation of the MFP, however : there is no millisecond conversion (yet), so the temporal resolution has been disabled, with full knowledge of the fact that this WILL cause anti-shadows in the Unity binding ! This is something to fix on the Unity side.

NOTE : the Visual Studio compiler throws a compile error if we do not suppress an assertion on matching vector allocators, because of how the chronology handles SetPairs. This did not pose any issues with GCC, so it is assumed that suppressing the error is safe for now.