Buggy proof-of-concept for DLL render
- Corrected wrong-order initialization of events
- Added clear method
- Added the renderCommand method, but...

It takes what is supposed to be an array pointer as its argument. But filling that array...doesn't work. It should : the memory was allocated on the C# side and there is no reason for anything to go wrong, but...it does.  Though no errors are thrown, going back to C# gives garbage values.

So instead, renderCommand which SHOULD return void and only initialize the array does nothing with that array and returns only the first event of the set that was pulled. Just to show that the conversion of a single event works.

So in other words, all that is left is to understand how the array works, and the first workable version of the DLL will come to life.