Actual array filling : "first" version of DLL, but...?
...the last event is missing.

In detail : function now returns void at it should. "processNoteData" was changed into an idiomatic conversion operator and moved to the noteData (and commandData for good measure) struct definition.

Now for the heart of the problem : this code does not seem to return the last element of the vector returned by performer.render(). Why ? I don't know !

I have tried everything under the sun to make this work : most notably, the awkward index management is removed, the code no longer writes any 0s instead relying on the fact that C# already initializes its arrays to all 0s (which it does). As well, the size is no longer needed, since we will never go past 4096 anyway ; if we did, we have bigger problems.

But even with the code stripped down to such a simple level, and though it seems to work on this side, tests show that the array is always left without the final element of the vector.

Actually, no, there is one case where that isn't true : if the vector has a single element, then that element is correctly translated into the array !!!!!!

This makes NO SENSE. The only point of failure is the enhanced for-loop, but the vector is properly sized (as the rest of the lib, but also prints confirm).

I have no clue what's happening. For now, I'm pushing this.