Fixed array-filling bug by changing unsigned long to uint64_t
Or : another example of why you should never use anything other than unambiguously sized stdint types...

In theory, because this DLL is compiled to x64, C++ *should* make unsigned longs 64-bit long. But it didn't. So in treating the 64-bit-pointing ulong pointer as a 32-bit unsigned long pointer, it wrote to each slot twice, missing every other event and corrupting them in the meantime by writing to both the lower and upper half.

I don't know why : again, unsigned long and uint64_t should be the same size for x64. It's because of this weirdness that stdint types should always be used.